update ffmpeg
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@32290 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/minidlna/ffmpeg-3.1/libavcodec/eatqi.c
@@ -94,7 +94,7 @@ static inline void tqi_idct_put(AVCodecContext *avctx, AVFrame *frame,
                                 int16_t (*block)[64])
 {
     TqiContext *t = avctx->priv_data;
-    int linesize = frame->linesize[0];
+    ptrdiff_t linesize = frame->linesize[0];
     uint8_t *dest_y  = frame->data[0] + t->mb_y * 16 * linesize           + t->mb_x * 16;
     uint8_t *dest_cb = frame->data[1] + t->mb_y *  8 * frame->linesize[1] + t->mb_x *  8;
     uint8_t *dest_cr = frame->data[2] + t->mb_y *  8 * frame->linesize[2] + t->mb_x *  8;
@@ -112,7 +112,7 @@ static inline void tqi_idct_put(AVCodecContext *avctx, AVFrame *frame,
 
 static void tqi_calculate_qtable(TqiContext *t, int quant)
 {
-    const int qscale = (215 - 2*quant)*5;
+    const int64_t qscale = (215 - 2*quant)*5;
     int i;
 
     t->intra_matrix[0] = (ff_inv_aanscales[0] * ff_mpeg1_default_intra_matrix[0]) >> 11;
